requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-dotenv
PyJWT
requests
beautifulsoup4
